+++ src/page/Works/WorkDetailPage.tsx
@@ -4,7 +4,7 @@ import { WorkBottomButton } from '@/components/Works/WorkDetail/WorkBottomButton
 import { WorkDetailMain } from '@/components/Works/WorkDetail/WorkDetailMain';
 import { WorkHeader } from '@/components/Works/WorkDetail/WorkHeader';
 import { WorkInfo } from '@/components/Works/WorkDetail/WorkInfo';
-import { RecruitmentDetailResponse } from '@/type/Work';
+import { RecruitmentDetailResponse } from '@/types/Work';
 import axios from 'axios';
 import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
